Test struct field types

+++ ctest/testcrate/src/t2.rs
@@ -3,8 +3,8 @@ pub type T2Bar = u32;
 
 #[repr(C)]
 pub struct T2Baz {
-    pub a: u8,
-    pub b: i32,
+    pub a: i64,
+    pub b: u32,
 }
 
 pub const T2C: i32 = 5;
